+++ packages/mui-material/src/utils/useRovingTabIndex.js
@@ -1,0 +1,4 @@
+'use client';
+import useRovingTabIndex from '@mui/utils/useRovingTabIndex';
+
+export default useRovingTabIndex;
