+++ packages/mui-material/src/utils/useRovingTabindex.ts
@@ -1,0 +1,3 @@
+import useRovingTabIndex from '@mui/utils/useRovingTabIndex';
+
+export default useRovingTabIndex;
