+++ packages/mui-utils/src/index.ts
@@ -49,4 +49,5 @@ export { default as unstable_resolveComponentProps } from './resolveComponentPro
 export { default as unstable_extractEventHandlers } from './extractEventHandlers';
 export { default as unstable_getReactNodeRef } from './getReactNodeRef';
 export { default as unstable_getReactElementRef } from './getReactElementRef';
+export { default as unstable_useRovingTabIndex } from './useRovingTabIndex';
 export * from './types';
