+++ packages/mui-utils/src/useRovingTabIndex/index.ts
@@ -1,0 +1,1 @@
+export { default } from './useRovingTabIndex';
